move utxo.rb to under wallet
