+++ lib/bitcoin/wallet.rb
@@ -4,5 +4,6 @@ module Bitcoin
     autoload :Account, 'bitcoin/wallet/account'
     autoload :DB, 'bitcoin/wallet/db'
     autoload :MasterKey, 'bitcoin/wallet/master_key'
+    autoload :Utxo, 'bitcoin/wallet/utxo'
   end
 end
